Split and enabled test_basic
