+++ python/ray/tests/BUILD
@@ -37,6 +37,11 @@ py_test_module_list(
     "test_basic_2.py",
     "test_basic_4.py",
     "test_basic_5.py",
+    "test_basic_6.py",
+    "test_basic_7.py",
+    "test_basic_8.py",
+    "test_basic_9.py",
+    "test_basic_10.py",
     "test_cancel.py",
     "test_gcs_fault_tolerance.py",
     "test_metrics_agent.py",
